Change Vector2.__repr__ to make it nicer.
Change the `repr` for `Vector2` objects.
If both `x` and `y` are set, and they're both
integers, it's more pleasant to only see those values,
even if some/all of the other values are cached.
This makes `repr(vector2_zero)` a lot easier to read.

Just noticed, we're not at 100% coverage right now!
Gotta fix that before next release.